♻️ refactor: KafkaTemplate 의존 제거 및 아키텍처 구조에 맞게 클래스 배치 수정
