+++ src/main/kotlin/com/uahannam/common/annotation/MessagingAdapter.kt
@@ -1,0 +1,8 @@
+package com.uahannam.common.annotation
+
+import org.springframework.stereotype.Component
+
+@Component
+@Target(AnnotationTarget.CLASS)
+@Retention(AnnotationRetention.RUNTIME)
+annotation class MessagingAdapter()
