+++ src/main/kotlin/com/uahannam/common/config/KafkaProducerConfig.kt
@@ -1,7 +1,7 @@
 package com.uahannam.common.config
 
-import com.uahannam.order.adapter.out.kafka.produce.dto.ModifyOrderStatusKafkaDto
-import com.uahannam.order.adapter.out.kafka.produce.dto.SaveOrderKafkaDto
+import com.uahannam.order.adapter.out.messaging.produce.dto.ModifyOrderStatusKafkaDto
+import com.uahannam.order.adapter.out.messaging.produce.dto.SaveOrderKafkaDto
 import org.apache.kafka.clients.producer.ProducerConfig
 import org.apache.kafka.common.serialization.StringSerializer
 import org.springframework.context.annotation.Bean
