+++ src/main/kotlin/com/uahannam/common/dto/base/BaseResponse.kt
@@ -1,4 +1,4 @@
-package com.uahannam.common.dto
+package com.uahannam.common.dto.base
 
 data class BaseResponse<T>(
     val data: T
