+++ src/main/kotlin/com/uahannam/order/application/service/CreateOrderService.kt
@@ -3,7 +3,7 @@ package com.uahannam.order.application.service
 import com.uahannam.common.annotation.UseCase
 import com.uahannam.order.application.port.`in`.model.CreateOrderCommand
 import com.uahannam.order.application.port.`in`.usecase.CreateOrderUseCase
-import com.uahannam.order.application.port.out.CreateOrderPort
+import com.uahannam.order.application.port.out.persistence.CreateOrderPort
 import org.springframework.transaction.annotation.Transactional
 
 @UseCase
