+++ src/main/kotlin/com/uahannam/order/application/service/LoadOrderService.kt
@@ -1,7 +1,7 @@
 package com.uahannam.order.application.service
 
 import com.uahannam.order.application.port.`in`.usecase.LoadOrderUseCase
-import com.uahannam.order.application.port.out.LoadOrderPort
+import com.uahannam.order.application.port.out.persistence.LoadOrderPort
 import org.springframework.stereotype.Service
 import org.springframework.transaction.annotation.Transactional
